Выполняет 4 задание "Модуляция"
